Added updating entity collections
This allows you to create entity collections using a filter function so that they will keep track of created and destroyed units.

This gives a big speed improvement for anything which needs to scan all the entities in the game.  See Gambia River for a demonstration of the lag with the old functionality.